refactor(simconnect): split object.go into AI aircraft and control files (#33)
Split object.go (340 lines) into 2 files:
- object.go: Generic object operations (AICreateSimulatedObject, AIReleaseControl, AIRemoveObject, EnumerateSimObjectsAndLiveries) + EX1 variants
- object-ai.go: AI aircraft creation (AICreateEnrouteATCAircraft, AICreateNonATCAircraft, AICreateParkedATCAircraft, AISetAircraftFlightPlan) + EX1 variants

No behavior changes — pure file reorganization.
